Split functions into files
